Fix segfault by matching base with the go snap
Since we stage the go snap it expects to use libc from its
base snap which is core20 - so if we are not base core20
ourself then these libs don't exist and it fails to run,
and hence segfaults. So switch to base core20 ourself and
don't stage libc6 as this can now be gotten from our own
base snap.

Signed-off-by: Alex Murray <murray.alex@gmail.com>